Implement KdTree#draw()
Use red-black trees to hold all existing line segments.
Create KdTreeDrawTest.java for easier UAT for draw().
In production code, we would create a facade to access StdDraw and
use dependency injection to mock the calls to StdDraw. However, this is
not possible because we cannot modify the KdTree public API.